Moving TextStats etc into separate file
